Move push rule ids extensions to vector module
